Removed some legacy-things and merged the two interface-files.
